backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
aiosqlite==0.20.0

# Data validation
pydantic==2.9.0
pydantic-settings==2.5.0

# NLP libraries (from original project)
pyenchant==3.2.2
nltk==3.9.1
fugashi==1.3.2
unidic-lite
ass==0.5.4
numpy

# Development
pytest==8.3.0
